Move all checkout complete logic to checkout_complete file.
