+++ saleor/order/models.py
@@ -29,6 +29,10 @@ from . import FulfillmentStatus, OrderEvents, OrderStatus
 
 
 class OrderQueryset(models.QuerySet):
+    def get_by_checkout_token(self, token):
+        """Return non-draft order with matched checkout token."""
+        return self.confirmed().filter(checkout_token=token).first()
+
     def confirmed(self):
         """Return non-draft orders."""
         return self.exclude(status=OrderStatus.DRAFT)
